api/files: Refactored file service (split into 2 services + assertions)
